Exception on calling creator.toolbox.addItems() with erasing old items fix #5067
