+++ packages/survey-creator-core/src/toolbox.ts
@@ -79,6 +79,9 @@ export class QuestionToolboxCategory extends Base {
 export class QuestionToolboxItem extends Action implements IQuestionToolboxItem {
   constructor(private item: IQuestionToolboxItem) {
     super(item);
+    if(!this.id) {
+      this.id = this.name;
+    }
   }
   className: string;
   iconName: string;
