+++ packages/survey-creator-core/tests/creator-base.tests.ts
@@ -36,7 +36,6 @@ import { TabJsonEditorTextareaPlugin } from "../src/components/tabs/json-editor-
 import { TabJsonEditorAcePlugin } from "../src/components/tabs/json-editor-ace";
 import { isTextInput } from "../src/creator-base";
 import { ItemValueWrapperViewModel } from "../src/components/item-value";
-import { ToolboxToolViewModel } from "../src/components/toolbox/toolbox-tool";
 
 import {
   getElementWrapperComponentData,
@@ -102,25 +101,6 @@ test("Click on toolbox on empty survey", (): any => {
   expect(creator.survey.pages[0].elements).toHaveLength(1);
   expect(creator.selectedElementName).toEqual("question1");
   expect(creator.selectedElement.getType()).toEqual("text");
-});
-test("Click on toolbox and cancel survey.lazyRendering", (): any => {
-  const creator = new CreatorTester();
-  expect(creator.survey.isLazyRendering).toEqual(true);
-  creator.clickToolboxItem({ type: "text" });
-  expect(creator.survey.isLazyRendering).toEqual(false);
-});
-test("Click on toolbox and insert into correct index", (): any => {
-  const creator = new CreatorTester();
-  creator.JSON = {
-    elements: [
-      { type: "text", name: "question1" },
-      { type: "text", name: "question2" }
-    ]
-  };
-  creator.selectElement(creator.survey.getQuestionByName("question1"));
-  creator.clickToolboxItem({ type: "text" });
-  expect(creator.selectedElementName).toEqual("question3");
-  expect(creator.survey.currentPage.elements[1].name).toEqual("question3");
 });
 test("Update JSON before drag&drop", (): any => {
   const creator = new CreatorTester();
@@ -1191,26 +1171,6 @@ test("Check page actions for pageEditMode is 'single'", (): any => {
   expect(pageModel.getActionById("duplicate").visible).toBeFalsy();
   expect(pageModel.getActionById("settings").visible).toBeTruthy();
 });
-test("Convert checkbox into rating", (): any => {
-  const creator = new CreatorTester();
-  creator.JSON = {
-    elements: [{ type: "checkbox", name: "question1", choices: [1, 2, 3, 4] }]
-  };
-  let q = creator.survey.getQuestionByName("question1");
-  creator.selectElement(q);
-  creator.convertCurrentQuestion("rating");
-  let el = <QuestionRatingModel>creator.selectedElement;
-  expect(el.getType()).toEqual("rating");
-  expect(el.rateValues).toHaveLength(4);
-  expect(el.rateValues[0].value).toEqual(1);
-  creator.clickToolboxItem(creator.toolbox.getItemByName("checkbox").json);
-  expect(creator.selectedElement.getType()).toEqual("checkbox");
-  creator.convertCurrentQuestion("rating");
-  el = <QuestionRatingModel>creator.selectedElement;
-  expect(el.getType()).toEqual("rating");
-  expect(el.rateValues).toHaveLength(0);
-});
-
 test("Convert text question into dropdown", (): any => {
   var creator = new CreatorTester();
   creator.JSON = {
@@ -1337,18 +1297,6 @@ test("Add question with default choices", (): any => {
   const question = <QuestionRadiogroupModel>survey.getAllQuestions()[0];
   expect(question.getType()).toEqual("radiogroup");
   expect(question.visibleChoices.length).toEqual(6);
-});
-test("Add question based on json in toolbox", (): any => {
-  const creator = new CreatorTester();
-  const toolboxItem = creator.toolbox.getItemByName("text");
-  toolboxItem.json.placeholder = "Test holder";
-  const survey: SurveyModel = creator.survey;
-  creator.currentAddQuestionType = "text";
-  creator.addNewQuestionInPage(() => { });
-  const question = <QuestionTextModel>survey.getAllQuestions()[0];
-  expect(question.getType()).toEqual("text");
-  expect(question.placeholder).toEqual("Test holder");
-  delete toolboxItem.json.placeholder;
 });
 test("getElementWrapperComponentName", (): any => {
   expect(getElementWrapperComponentName(null, "logo-image", false)).toEqual("svc-logo-image");
@@ -2260,43 +2208,6 @@ test("ConvertTo & addNewQuestion for panel & maxNestedPanels ", (): any => {
   expect(creator.getAvailableToolboxItems()).toHaveLength(itemCount);
 });
 
-test("Has one item type in convertTo", (): any => {
-  CustomWidgetCollection.Instance.add({
-    name: "text",
-    title: "Single-Line Input",
-    widgetIsLoaded: () => { return true; },
-    isFit: () => (question) => { return question.getType() === "text"; }
-  }, "customtype");
-  const creator = new CreatorTester();
-  creator.JSON = {
-    elements: [
-      { type: "radiogroup", name: "q1" }
-    ]
-  };
-  let counter = 0;
-  creator.toolbox.itemNames.forEach(item => { if (item === "text") counter++; });
-  expect(counter).toEqual(1);
-  const question = creator.survey.getQuestionByName("q1");
-  creator.selectElement(question);
-
-  const questionModel = new QuestionAdornerViewModel(
-    creator,
-    question,
-    undefined
-  );
-  const items = questionModel.getConvertToTypesActions();
-  const popup = questionModel.getActionById("convertTo").popupModel;
-  expect(popup).toBeTruthy();
-  const list = popup.contentComponentData.model;
-  expect(list).toBeTruthy();
-  counter = 0;
-  list.actions.forEach(item => {
-    if (item.id === "text") counter++;
-  });
-  expect(counter).toEqual(1);
-  CustomWidgetCollection.Instance.clear();
-});
-
 test("ConverTo, change title of question item", (): any => {
   const creator = new CreatorTester();
   creator.toolbox.getItemByName("radiogroup").title = "Single selector";
@@ -2316,52 +2227,6 @@ test("ConverTo, change title of question item", (): any => {
   );
   const action = questionModel.getActionById("convertTo");
   expect(action.title).toEqual("Single selector");
-});
-test("ConvertTo, show custom widgets in ConvertTo action", (): any => {
-  const widget = {
-    name: "test_widget",
-    title: "Test Widget",
-    iconName: "icon-editor",
-    widgetIsLoaded: function () {
-      return true;
-    },
-    isFit: function (question) {
-      return question.getType() === "test_widget";
-    },
-    init() {
-      //Register a new type using the empty question as the base.
-      Serializer.addClass("test_widget", [], null, "empty");
-    },
-    htmlTemplate:
-      "<div>This is test widget</div>",
-    afterRender: function (question, element) {
-    }
-  };
-
-  CustomWidgetCollection.Instance.add(widget, "customtype");
-
-  const creator = new CreatorTester({ questionTypes: ["text", "comment"] });
-  creator.JSON = {
-    elements: [
-      { type: "text", name: "q1" }
-    ]
-  };
-  expect(creator.toolbox.items).toHaveLength(3);
-  const question = creator.survey.getQuestionByName("q1");
-  creator.selectElement(question);
-
-  const questionModel = new QuestionAdornerViewModel(
-    creator,
-    question,
-    undefined
-  );
-  const items = questionModel.getConvertToTypesActions();
-  expect(items).toHaveLength(3);
-  expect(items[0].id).toEqual("text");
-  expect(items[1].id).toEqual("comment");
-  expect(items[2].id).toEqual("test_widget");
-  expect(items[2].iconName).toEqual("icon-editor");
-  CustomWidgetCollection.Instance.clear();
 });
 
 test("ConvertTo separators", (): any => {
@@ -3926,45 +3791,6 @@ test("Reordering on drag&drop", (): any => {
   expect(creator.undoRedoManager.canRedo()).toBeTruthy();
 });
 
-test("Reason of question Added from toolbox, onclicking add question button, on duplicated question, panel, page", (): any => {
-  const creator = new CreatorTester();
-  const reason = [];
-  creator.onQuestionAdded.add(function (sender, options) {
-    reason.push(options.reason);
-  });
-
-  creator.clickToolboxItem({ type: "text" });
-  expect(reason).toHaveLength(1);
-  expect(reason[0]).toEqual("ADDED_FROM_TOOLBOX");
-
-  creator.fastCopyQuestion(creator.survey.getAllQuestions()[0]);
-  expect(reason).toHaveLength(2);
-  expect(reason[1]).toEqual("ELEMENT_COPIED");
-
-  const pageAdorner = new PageAdorner(creator, creator.survey.pages[0]);
-  pageAdorner.addNewQuestion(pageAdorner, undefined);
-  expect(reason).toHaveLength(3);
-  expect(reason[2]).toEqual("ADDED_FROM_PAGEBUTTON");
-
-  creator.copyPage(creator.survey.pages[0]);
-  expect(reason).toHaveLength(6);
-  expect(reason[3]).toEqual("ELEMENT_COPIED");
-  expect(reason[4]).toEqual("ELEMENT_COPIED");
-  expect(reason[5]).toEqual("ELEMENT_COPIED");
-
-  const toolboxViwer = new ToolboxToolViewModel(creator.toolbox.items[0], creator);
-  toolboxViwer.click({});
-  expect(reason).toHaveLength(7);
-  expect(reason[6]).toEqual("ADDED_FROM_TOOLBOX");
-  creator.onDragDropItemStart();
-  creator.survey.pages[0].addNewQuestion("text", "qqq1");
-  expect(reason).toHaveLength(8);
-  expect(reason[7]).toEqual("DROPPED_FROM_TOOLBOX");
-  expect(creator.selectedElementName).toBe("question7");
-  creator.convertCurrentQuestion("comment");
-  expect(reason).toHaveLength(9);
-  expect(reason[8]).toEqual("ELEMENT_CONVERTED");
-});
 test("Initial Property Grid category expanded state", (): any => {
   const creator = new CreatorTester();
   let survey: SurveyModel;
@@ -4377,17 +4203,4 @@ test("New ghost page shouldn't be created if onPageAdding sets allow to false", 
   expect(creator.survey.pages).toHaveLength(1);
   expect(desigerTab.newPage).toBeFalsy();
   expect(desigerTab.showNewPage).toBeFalsy();
-});
-test("Add-remove toolbox items, #5271", (): any => {
-  const creator = new CreatorTester();
-  creator.toolbox.allowExpandMultipleCategories = true;
-  creator.toolbox.showCategoryTitles = true;
-  creator.onQuestionAdded.add((sender, options) => {
-    if (options.question.getType() === "dropdown") {
-      creator.toolbox.removeItem("dropdown");
-    }
-  });
-  creator.currentAddQuestionType = "dropdown";
-  creator.addNewQuestionInPage(() => {}, undefined, "dropdown");
-  expect(creator.getAddNewQuestionText()).toEqual("Add Question");
 });
